+++ src/layout/components/index.ts
@@ -1,5 +1,6 @@
 import { packEnclose, packSiblings } from 'd3-hierarchy'
-import { Node, Edge, connectedComponents } from '../..'
+import { connectedComponents } from '../../utils/api'
+import { Node, Edge } from '../../types'
 
 export type Options = Partial<{
   padding: number
