+++ src/layout/force/index.ts
@@ -11,8 +11,7 @@ import {
   SimulationLinkDatum,
   SimulationNodeDatum
 } from 'd3-force'
-import { Extend } from '../../types'
-import { Node, Edge } from '../..'
+import { Node, Edge, Extend } from '../../types'
 
 export type Options = Partial<{
   nodeStrength: number
