+++ src/types/internal.ts
@@ -1,0 +1,3 @@
+export type Extend<T, R> = {
+  [K in Exclude<keyof T, keyof R>]: T[K]
+} & R
